Save text entry to database

+++ app/src/main/java/me/partypronl/quibble/entry/JournalColumnEntry.kt
@@ -14,7 +14,7 @@ class JournalColumnEntry: JournalEntryType(
 
     }
 
-    override suspend fun getCardFromJournalEntryId(journalEntryId: Int): @Composable (() -> Unit) {
+    override suspend fun getCardFromJournalEntryId(journalEntryId: Long): @Composable (() -> Unit) {
         return {}
     }
 }
